update load textfile

+++ image_filename.py
@@ -1,14 +1,14 @@
-from folder_paths import base_path
-import os, torch, json
-from PIL import Image, ImageOps
+import os, json, random
+from PIL import Image
 from PIL.PngImagePlugin import PngInfo
 import numpy as np
-import hashlib
+from typing import Any
 from nodes import SaveImage
 
 from comfy.cli_args import args
 import folder_paths
-
+from comfy_api.latest import io
+from pathlib import Path
 
 class SaveFilename(SaveImage):
     CATEGORY = "quicknodes/images"
@@ -22,34 +22,40 @@ class SaveFilename(SaveImage):
         outfile = ret['ui']['images'][0]
         directory = folder_paths.get_output_directory()
         if 'subfolder' in outfile and outfile['subfolder']: directory = os.path.join(directory, outfile['subfolder'])
-        ret['result'] = ( os.path.join(directory, outfile['filename']), )
+        ret['result'] = ( os.path.join(directory, outfile['filename']), ) # type: ignore
         return ret
     
-    
-class LoadTextfile:
-    CATEGORY = "quicknodes/images"
-
-    RETURN_TYPES = ("STRING",)
-    FUNCTION = "func"
-    OUTPUT_NODE = True
-
+class LoadTextfile(io.ComfyNode):
     @classmethod
-    def INPUT_TYPES(cls):
-        return {
-            "required": {
-                "filepath": ("STRING", {"default": "output", "tooltip": "The filepath, extension will be changed."}),
-                "remove": ("BOOLEAN", {"default": False, "tooltip": "Remove the file after reading."}),
-            }
-        }
-
-    def func(self, filepath, remove ):
-        filepath = os.path.splitext(filepath)[0]+".txt"
-        with open(filepath, 'r', encoding="utf-8") as fh: s = fh.read()
-        if remove: os.remove(filepath)
-        return (s.strip(),)
+    def define_schema(cls):
+        return io.Schema(
+            node_id      = "LoadTextfile",
+            display_name = "Load Textfile",
+            category     = "quicknodes/io",
+            inputs       = [ 
+                io.String.Input("filepath", default="output", tooltip="The filepath, extension will be changed."),
+                io.Boolean.Input("remove", default=False, tooltip="Remove the file after reading."),
+            ],
+            outputs      = [ io.String.Output("text"), ],
+        )
+    
+    @classmethod
+    def execute(cls, filepath, remove ): # type: ignore
+        filepath = Path(os.path.splitext(filepath)[0]+".txt")
+        if filepath.exists():
+            with open(filepath, 'r', encoding="utf-8") as fh: s = fh.read()
+            if remove: os.remove(filepath)
+            return io.NodeOutput(s.strip(),)
+        else:
+            print(f"File not found: {filepath}")
+            return io.NodeOutput("")
+    
+    @classmethod
+    def fingerprint_inputs(cls, **kwargs) -> Any:
+        return random.random()
     
 class SaveTextfile:
-    CATEGORY = "quicknodes/images"
+    CATEGORY = "quicknodes/io"
 
     RETURN_TYPES = ()
     FUNCTION = "func"
@@ -91,7 +97,7 @@ class SaveImageExplicitPath:
     RETURN_TYPES = ()
     FUNCTION = "save_images"
     OUTPUT_NODE = True
-    CATEGORY = "quicknodes"
+    CATEGORY = "quicknodes/io"
 
     def save_images(self, images, filepath, compress, prompt=None, extra_pnginfo=None):
         basepath, extension = os.path.splitext(filepath)
